Minor improvements.
* Fixes inefficiency where a new Array was allocated multiple
times, a result of Java invading my brain. This should not have
resulted in any incorrect behavior, just inefficient.
* Replaces some for-ins with contains(element:). I originally
made them for-ins for a stupid reason (to match the other overload
which is cleaner as a for-in than contains(where:)), now I see "that"
isn't worth someone wondering why I didn't just use contains(element:).
This change should just be syntactic sugar since both operations were
linear to begin with, this should not have resulted in any incorrect
behavior either.

On branch master

Changes to be committed:
	modified:   RelativeView/Ancestor.swift
	modified:   RelativeView/Descendant.swift
	modified:   RelativeView/Sibling.swift
	modified:   relative-view.podspec